decimal coercion was only working for input datatype string
However, I think it should work for numbers and "number stringified"
this issue was noticed while using reitit. The following scenario:

initial setup: build a `decimal?` validation spec and performing the
following experiment:

1. using the spec in a route using `:form` parameters
2. using the spec in a route using `:body` parameters

The 1. works just fine as expected and 2. doesn't. And the reason is
related with `:form` parameters reading all the inputs as strings by
definition while the body of a json file maintains the numeric
datatypes.

This following PR attempts to fix this issue by coercion numeric
datatypes to bigdecimals too.